+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Unit test checking database reinitialization when no parameters passed
+
 ## [0.2.1] - 2021-03-07
 
 ### Added